disable snacks dashboard rename auto-session
