+++ init.lua
@@ -130,7 +130,7 @@ require('lazy').setup({
       -- or leave it empty to use the default settings
       -- refer to the configuration section below
       bigfile = { enabled = true },
-      dashboard = { enabled = true },
+      dashboard = { enabled = false },
       notifier = {
         enabled = true,
         timeout = 3000,
@@ -924,7 +924,7 @@ require('lazy').setup({
   require 'plugins.gitsigns', -- adds gitsigns recommend keymaps
   require 'plugins.codeium',
   require 'plugins.chatgpt',
-  require 'plugins.auto_session',
+  require 'plugins.auto-session',
 
   -- NOTE: The import below can automatically add your own plugins, configuration, etc from `lua/custom/plugins/*.lua`
   --    This is the easiest way to modularize your config.
